### 2021-08-18 #3629 @whedon
[PRE REVIEW]: TX2: Transformer eXplainability and eXploration
**Submitting author:** @WildfireXIII (<a href="http://orcid.org/0000-0002-5036-5433">Nathan Martindale</a>)
**Repository:** <a href="https://github.com/ORNL/tx2" target ="_blank">https://github.com/ORNL/tx2</a>
**Version:** 2021.8.0
