### 2021-08-19 #3629 @fabian-s
hey there, @richardlitt @desilinguist @assenmacher-mat
would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
