### 2021-08-19 #3629 @RichardLitt
Can’t this time. Thanks for asking.

On Thursday, August 19, 2021, Fabian Scheipl ***@***.***>
wrote:

> hey there, @RichardLitt <https://github.com/RichardLitt> @desilinguist
> <https://github.com/desilinguist> @assenmacher-mat
> <https://github.com/assenmacher-mat>
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
