### 2021-08-24 #3652 @whedon
[REVIEW]: TX2: Transformer eXplainability and eXploration
**Submitting author:** @WildfireXIII (<a href="http://orcid.org/0000-0002-5036-5433">Nathan Martindale</a>)
**Repository:** <a href="https://github.com/ORNL/tx2" target ="_blank">https://github.com/ORNL/tx2</a>
**Version:** v1.0.0
