### 2021-09-08 #3652 @assenmacher-mat
Two small issues:
- Maybe there should also be an explicit manual on how exactly install pytoch w/o gpu
- Regarding the point I haven't marked (see above): I could not find any comment regarding this
