### 2021-09-16 #3652 @fabian-s
@WildfireXIII  : please add a CONTRIBUTING.md file or similar to your repo, as pointed out by @assenmacher-mat
