### 2021-09-16 453f125
Update README with non-gpu torch instructions and JOSS badge
