### 2021-09-16 #3652 @stewartsl
> @WildfireXIII : please add a CONTRIBUTING.md file or similar to your repo, as pointed out by @assenmacher-mat

@fabian-s please note that there is currently a contributing.MD file in master [tx2 contributing](https://github.com/ORNL/tx2/blob/master/CONTRIBUTING.md). does this sufficiently address your concern?
