### 2021-09-16 #3652 @fabian-s
> @fabian-s please note that there is currently a contributing.MD file in master [tx2 contributing](https://github.com/ORNL/tx2/blob/master/CONTRIBUTING.md). does this sufficiently address your concern?

sure, that seems sufficient.
