### 2021-10-29 ORNL/tx2#1 @deniederhut
Paper suggestions / questions
- In the summary, instead of saying the library "assumes the use" of HuggingFace, I might say that in that first sentence that it's built for understanding the behavior of HuggingFace transformer models
- In the introduction, you mention the Attention is not Explanation paper (awesome), but this method is attention-based, no?
- I think we're missing a reference to the Vaswani 2017 paper
- I was expecting to see a reference to https://github.com/jessevig/bertviz here -- is there a reason it wasn't included?
- I tend to prefer journal references over arxiv when available. Leland lists JOSS in the "how to cite" section for UMAP (it was also presented at SciPy the same year https://www.youtube.com/watch?v=nq6iPZVUxZU, but alas no paper)
- You can find guidance on how to cite Jupyter here: https://github.com/jupyter/jupyter/issues/190
- scikit-learn also has guidance for citations https://scikit-learn.org/stable/about.html#citing-scikit-learn
- I don't know that I would agree with "These users will also tend to perform some or most of their work within the Jupyter ecosystem" -- there is a rather famous talk about this by a researcher at Allen. I don't think this is a dealbreaker or anything; Jupyter is wildly popular and facilitates interactive exploration, even if you wouldn't have used Jupyter to build the model in the first place.
- On that note, what if I have a fully trained model already? Can I load in a pretrained head? What if the task isn't classification, but regression or a sequence-to-sequence?
