### 2021-10-31 #3652 @fabian-s
> @fabian-s what's the protocol for me, once the linked issues are addressed and eventually closed, do I throw a comment up for it here?

sure, that will work. if you reference `openjournals/joss-reviews#3652` in your comments in the linked issues (or commit messages), notifications will pop up here as well.
