### 2021-11-09 #3652 @deniederhut
Hey @fabian-s , could I ask you to weigh in on https://github.com/ORNL/tx2/issues/4? I'm not super clear on the testing requirements for JOSS submissions.
