### 2021-11-09 ORNL/tx2#4 @fabian-s
It's sub-par but acceptable, I'd say. The examples provide a manual way to check that the tool is working as expected. Obviously proper unit or regression test and adding CI to the repo would be preferable

@deniederhut would you say that these two notebooks cover TX21's central functionality?
If so, I'd recommend to cross that item off the checklist.
If not, please identify what's missing, and then we can unblock once @WildfireXIII has added examples for those parts.
