### 2021-11-17 ORNL/tx2#4 @deniederhut
Thanks for chiming in! I've tried that first notebook and it's failing with an error. I've filed a separate ticket about it, so I think we can close this one.
