### 2021-11-22 #3652 @stewartsl
> @WildfireXIII please update us on your progress in adressing these issues.

We have made some progress and will continue to work on these next week after the holiday break.
