### 2021-11-29 642277f
Fix conda and pip install lines in README
