### 2021-11-29 6d06c43
Update paper with better references and back off jupyter claim
