+++ paper/paper.bib
@@ -22,9 +22,17 @@
 }
 
 
-@misc{jupyternotebook,
-	title = {Jupyter Notebook},
-	howpublished = {\url{https://github.com/jupyter/notebook}},
+@inproceedings{jupyternotebook,
+       booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+          editor = {Fernando Loizides and Birgit Scmidt},
+           title = {Jupyter Notebooks - a publishing format for reproducible computational workflows},
+          author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and  Jupyter development team},
+       publisher = {IOS Press},
+         address = {Netherlands},
+            year = {2016},
+           pages = {87--90},
+             url = {https://eprints.soton.ac.uk/403913/},
+        abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.}
 }
 
 
@@ -53,13 +61,14 @@
 }
 
 
-@misc{mcinnes2020umap,
-	title={UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction},
-	author={Leland McInnes and John Healy and James Melville},
-	year={2020},
-	eprint={1802.03426},
-	archivePrefix={arXiv},
-	primaryClass={stat.ML}
+@article{mcinnes2018umap,
+  title={UMAP: Uniform Manifold Approximation and Projection},
+  author={McInnes, Leland and Healy, John and Saul, Nathaniel and Grossberger, Lukas},
+  journal={The Journal of Open Source Software},
+  volume={3},
+  number={29},
+  pages={861},
+  year={2018}
 }
 
 
@@ -102,12 +111,27 @@
 }
 
 
-@misc{sklearn-clustering,
-	title = {sklearn.clustering},
-	howpublished = {\url{https://scikit-learn.org/stable/modules/clustering.html}},
+@article{scikit-learn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
 }
 
 @misc{black-linter,
 	title = {Black - The Uncompromising Code Formatter},
 	howpublished = {\url{https://github.com/psf/black}},
+}
+
+# Original transformers paper
+@article{vaswani2017attention,
+  title={Attention is all you need},
+  author={Vaswani, Ashish and Shazeer, Noam and Parmar, Niki and Uszkoreit, Jakob and Jones, Llion and Gomez, Aidan N and Kaiser, Lukasz and Polosukhin, Illia},
+  journal={arXiv preprint arXiv:1706.03762},
+  year={2017}
 }
