### 2021-11-29 ORNL/tx2#1 @WildfireXIII
> In the summary, instead of saying the library "assumes the use" of HuggingFace, I might say that in that first sentence that it's built for understanding the behavior of HuggingFace transformer models

This library doesn't technically need HuggingFace transformers specifically, it's a lot easier and we use them in the example notebooks, but any transformer model for which the user can implement the four interaction functions should technically work.

> In the introduction, you mention the Attention is not Explanation paper (awesome), but this method is attention-based, no?

It's not actually attention-based, our salience map approach is a bit more coarse than that - the "word importance" is calculated by removing each word of a text individually and measuring the change in the predicted output probabilities. (This is also what allows the approach to be relatively model agnostic.)

> I think we're missing a reference to the Vaswani 2017 paper

Oops, you're right! Fixed!

> I was expecting to see a reference to https://github.com/jessevig/bertviz here

We do reference it in the statement of need section, their repo asked to cite their paper instead, "A Multiscale Visualization of Attention in the Transformer Model", it's the (Vig, 2019) ref.

> I don't know that I would agree with "These users will also tend to perform some or most of their work within the Jupyter ecosystem"

That's fair, I backed off the claim.

> On that note, what if I have a fully trained model already? Can I load in a pretrained head? What if the task isn't classification, but regression or a sequence-to-sequence?

Yes, TX2 doesn't handle any of the training and makes no assumptions about how the training was done, so it's up to the user to provide a fully trained model to TX2. Unfortunately TX2 can only handle exploring the model from a classification task perspective. You could in principle have a model that was originally trained on a regression task, but for the interaction functions TX2 is expecting to make sense (`classify()` and `soft_classify()`), the model has to have some way to provide a classification output and classification probabilities.

Citations for UMAP, Jupyter, and scikit-learn are fixed!
