### 2021-11-29 a24ad8a
Add test skeleton and begin exploring using numpy inputs
