+++ requirements.txt
@@ -11,3 +11,4 @@ ipywidgets
 autodocsumm
 sphinx
 sphinx-rtd-theme
+pytest
