### 2021-11-30 5ca862d
Refactor visualizations to not require dataframes and implement tests
