### 2021-11-30 6540933
Add more visualization tests and fix bug with rowcount in plot_wordclouds
