### 2021-11-30 49dff94
Fix dataframe usage for calc and add tests
