+++ tx2/calc.py
@@ -1,24 +1,16 @@
 """Helper calculation functions for the wrapper and dashboard."""
 
-from nltk.corpus import stopwords
+from typing import Any, Dict, List, Tuple, Union
+
 import numpy as np
 import pandas as pd
-from sklearn.cluster import (
-    DBSCAN,
-    KMeans,
-    AffinityPropagation,
-    Birch,
-    OPTICS,
-    AgglomerativeClustering,
-    SpectralClustering,
-    SpectralBiclustering,
-    SpectralCoclustering,
-    MiniBatchKMeans,
-    FeatureAgglomeration,
-    MeanShift,
-)
+from nltk.corpus import stopwords
+from sklearn.cluster import (DBSCAN, OPTICS, AffinityPropagation,
+                             AgglomerativeClustering, Birch,
+                             FeatureAgglomeration, KMeans, MeanShift,
+                             MiniBatchKMeans, SpectralBiclustering,
+                             SpectralClustering, SpectralCoclustering)
 from sklearn.feature_extraction.text import CountVectorizer
-from typing import Dict, List, Tuple, Any
 
 from tx2 import utils
 
@@ -127,23 +119,19 @@ def aggregate_cluster_salience_maps(
 
 
 def frequent_words_in_cluster(
-    df: pd.DataFrame,
-    cluster_indices: List[int],
-    input_col_name: str,
+    texts: Union[np.ndarray, pd.Series],
 ) -> List[Tuple[str, int]]:
     """Finds the most frequently occurring words for each cluster given.
 
-    :param df: The dataframe containing the test data.
-    :param cluster_indices: The list of indices of the points in the desired cluster.
-    :param input_col_name: The name of the column in :code:`df` that contains the input text.
+    :param texts: The filtered list of texts to find the frequent words for
 
     :return: A list of tuples, each tuple containing the word and the number of times
         it appears in that cluster.
     """
     counter = CountVectorizer(stop_words=stopwords.words("english"))
-    cv_fit = counter.fit_transform(df.iloc[cluster_indices][input_col_name].values)
+    cv_fit = counter.fit_transform(texts)
     freq_words = sorted(
-        zip(counter.get_feature_names(), cv_fit.toarray().sum(axis=0)),
+        zip(counter.get_feature_names_out(), cv_fit.toarray().sum(axis=0)),
         key=lambda x: x[1],
         reverse=True,
     )
@@ -152,26 +140,20 @@ def frequent_words_in_cluster(
 
 
 def frequent_words_by_class_in_cluster(
-    df: pd.DataFrame,
     freq_words: List[Tuple[str, int]],
     encodings: Dict[str, int],
-    cluster_indices: List[int],
-    input_col_name: str,
-    classification_col_name: str,
+    cluster_texts: Union[np.ndarray, pd.Series],
+    cluster_text_labels: Union[np.ndarray, pd.Series],
 ) -> Dict[str, Dict[Any, int]]:
     """Takes the frequent words of a cluster and splits the counts up based on the
     classification of the entry they fall under. (This gives a rough distribution for
     what category the words fall under within the cluster.)
 
-    :param df: The dataframe containing the test data
     :param freq_words: An array of tuples of the words and their number of occurences,
         see :meth:`tx2.calc.frequent_words_in_cluster`
     :param encodings: The dictionary of class/category encodings.
-    :param cluster_indices: The list of indices of the points in the desired cluster.
-    :param classification_col_name: The name of the column in :code:`df` that contains
-        the target class value.
-    :param input_col_name: The name of the column in :code:`df` that contains the
-        input text.
+    :param cluster_texts: The set of texts from the desired cluster entries.
[…]
