### 2021-11-30 2ef287a
Remove dataframe usage from wrapper and add tests
