+++ tx2/wrapper.py
@@ -1,19 +1,18 @@
 """The wrapper class around a transformer and its functionality."""
 
 import logging
-import numpy as np
 import os
+from typing import Dict, List, Union
+
+import numpy as np
 import pandas as pd
 import torch
-
+import umap
 # TODO: not crazy about this, but library agnosticism later
 from torch.utils.data import DataLoader
-
 from tqdm.autonotebook import tqdm
-from typing import List, Dict
-import umap
 
-from tx2 import utils, calc, dataset
+from tx2 import calc, dataset, utils
 from tx2.cache import check, read, write
 
 
@@ -28,11 +27,11 @@ class Wrapper:
 
     def __init__(
         self,
-        train_df: pd.DataFrame,
-        test_df: pd.DataFrame,
+        train_texts: Union[np.ndarray, pd.Series],
+        train_labels: Union[np.ndarray, pd.Series],
+        test_texts: Union[np.ndarray, pd.Series],
+        test_labels: Union[np.ndarray, pd.Series],
         encodings: Dict[str, int],
-        input_col_name: str,
-        target_col_name: str,
         classifier=None,
         language_model=None,
         tokenizer=None,
@@ -41,12 +40,11 @@ class Wrapper:
     ):
         """Constructor.
 
-        :param train_df: A dataframe containing all training instances.
-        :param test_df: Dataframe containing test instances.
+        :param train_texts: A set of text entries that were used during the model's training process.
+        :param train_labels: The set of class labels for train_texts.
+        :param test_texts: The set of text entries that the model hadn't seen during training.
+        :param test_labels: The set of class labels for test_texts.
         :param encodings: A dictionary associating class label names with integer values.
-        :param input_col_name: The name of the column in both dataframes with the input text.
-        :param target_col_name: The name of the column in both dataframes with the classification
-            target.
         :param classifier: A class/network containing a language model and classification head.
             **Running this variable as a function by default should send the passed inputs through
             the entire network and return the argmaxed classification index (reverse encoding)**.
@@ -58,11 +56,15 @@ class Wrapper:
             pass it here. Note that **this argument is not required**, if the user intends to
             manually specify classification functions.
         :param tokenizer: A `huggingface tokenizer
-            <https://huggingface.co/transformers/main_classes/tokenizer.html>`_.
+            <https://huggingface.co/transformers/main_classes/tokenizer.html>`_. Note that **this
+            argument is not required**, if the user intends to manually specify encode and
+            classification functions.
         :param cache_path: The directory path to cache intermediate outputs from the
             :meth:`tx2.wrapper.Wrapper.prepare` function. This allows the wrapper to precompute
             needed values for the dashboard to reduce render time and allow rerunning all wrapper
-            code without needing to recompute.
+            code without needing to recompute. Note that every wrapper/dashboard instance is expected
+            to have a unique cache path, otherwise filenames will conflict. You will need to set
+            this if you intend to use more than one dashboard.
         :param overwrite: Whether to ignore the cache and overwrite previous results or not.
         """
 
@@ -98,14 +100,25 @@ class Wrapper:
         """A function to take a single text entry and return an encoded version of it. The default
         function will utilize the tokenizer given in the constructor if available."""
 
-        self.train_df = train_df.reset_index(drop=True)
-        """Dataframe containing all training data instances."""
-        self.test_df = test_df.reset_index(drop=True)
-        """Dataframe containing all testing data instances."""
[…]
