### 2021-12-01 a2e6b16
Fix dataframe usage in dashboard and add tests
