### 2021-12-06 ea8a8e9
Fix issue with debounce in vis tests and fix plot embedding graph
