### 2021-12-06 a203a87
Fix several more issues in new scatter function call
