### 2021-12-06 3b6522e
Fix plot embeddings error when no incorrect
