### 2021-12-06 ae7a860
Update README, notebooks, and setup
