### 2021-12-06 ORNL/tx2#4 @WildfireXIII
Updating this thread for awareness - we ended up adding some basic unit tests to aid in some of our updates, the test suite can be run via `pytest` from the root directory. (I've updated the README to include this as well.)
