+++ tests/test_wrapper.py
@@ -31,20 +31,21 @@ def test_wrapper_init_np_no_crash(dummy_np_data, dummy_encodings):
 # TODO: tests to ensure exceptions are thrown in prepare if incorrect combination of things have not been specified.
 
 
-def test_wrapper_prepare_no_crash(dummy_df, dummy_encodings, dummy_model, clear_files_teardown):
+def test_wrapper_prepare_no_crash(
+    dummy_df, dummy_encodings, dummy_model, clear_files_teardown
+):
     wrapper = Wrapper(
         train_texts=dummy_df.text,
         train_labels=dummy_df.target,
         test_texts=dummy_df.text,
         test_labels=dummy_df.target,
         encodings=dummy_encodings,
-        cache_path="testdata"
+        cache_path="testdata",
     )
-    
+
     wrapper.encode_function = dummy_model.custom_encode
     wrapper.classification_function = dummy_model.custom_classify
     wrapper.embedding_function = dummy_model.custom_embedding
     wrapper.soft_classification_function = dummy_model.custom_softclassify
-    
-    wrapper.prepare(umap_args=dict(n_neighbors=2))
 
+    wrapper.prepare(umap_args=dict(n_neighbors=2))
