+++ tx2/calc.py
@@ -5,11 +5,20 @@ from typing import Any, Dict, List, Tuple, Union
 import numpy as np
 import pandas as pd
 from nltk.corpus import stopwords
-from sklearn.cluster import (DBSCAN, OPTICS, AffinityPropagation,
-                             AgglomerativeClustering, Birch,
-                             FeatureAgglomeration, KMeans, MeanShift,
-                             MiniBatchKMeans, SpectralBiclustering,
-                             SpectralClustering, SpectralCoclustering)
+from sklearn.cluster import (
+    DBSCAN,
+    OPTICS,
+    AffinityPropagation,
+    AgglomerativeClustering,
+    Birch,
+    FeatureAgglomeration,
+    KMeans,
+    MeanShift,
+    MiniBatchKMeans,
+    SpectralBiclustering,
+    SpectralClustering,
+    SpectralCoclustering,
+)
 from sklearn.feature_extraction.text import CountVectorizer
 
 from tx2 import utils
