+++ tests/conftest.py
@@ -14,13 +14,17 @@ from tx2.wrapper import Wrapper
 def replacement_debounce():
     """This is so that we can replace utils.debounce and actually get
     error messages."""
+
     def undebounce(wait):
         def decorator(fn):
             def undebounced(*args, **kwargs):
                 print("WE'RE DOING IT LIVE")
                 fn(*args, **kwargs)
+
             return undebounced
+
         return decorator
+
     return undebounce
 
 
@@ -64,16 +68,15 @@ def dummy_encodings():
 
 @pytest.fixture
 def dummy_model(dummy_df):
-
     class model:
         def __init__(self, df):
-            self.vectorizer = CountVectorizer(stop_words='english')
-            
+            self.vectorizer = CountVectorizer(stop_words="english")
+
             x = self.vectorizer.fit_transform(df.text)
 
             self.clf = LogisticRegression()
             self.clf.fit(x, df.target)
-            
+
         def custom_encode(self, text):
             transformed = torch.tensor(self.vectorizer.transform([text]).toarray())
             return torch.squeeze(transformed)
@@ -86,11 +89,11 @@ def dummy_model(dummy_df):
 
         def custom_softclassify(self, inputs):
             return torch.tensor(self.clf.predict_proba(inputs))
-        
+
     return model(dummy_df)
 
 
-@pytest.fixture(scope='function')
+@pytest.fixture(scope="function")
 def dummy_wrapper(dummy_df, dummy_encodings, dummy_model, clear_files_teardown):
     wrapper = Wrapper(
         train_texts=dummy_df.text,
@@ -98,20 +101,20 @@ def dummy_wrapper(dummy_df, dummy_encodings, dummy_model, clear_files_teardown):
         test_texts=dummy_df.text,
         test_labels=dummy_df.target,
         encodings=dummy_encodings,
-        cache_path="testdata"
+        cache_path="testdata",
     )
-    
+
     wrapper.encode_function = dummy_model.custom_encode
     wrapper.classification_function = dummy_model.custom_classify
     wrapper.embedding_function = dummy_model.custom_embedding
     wrapper.soft_classification_function = dummy_model.custom_softclassify
-    
+
     wrapper.prepare(umap_args=dict(n_neighbors=2))
 
     return wrapper
-    
 
-@pytest.fixture(scope='session')
+
+@pytest.fixture(scope="session")
 def clear_files_teardown():
     yield None
     os.system("rm -rf testdata/")
