### 2021-12-08 #3652 @WildfireXIII
@deniederhut Hey! Just wanted to leave a note that I've made an update to each of the issues, so it's ready if you want to take another pass at it.
