### 2021-12-09 ORNL/tx2#1 @deniederhut
Cool, thanks for adding all those! Three more comments:

1. Each line is numbered. Is that on purpose?
2. I'm listed as the reviewer twice -- hopefully not because I was twice as slow? 😉
3. I don't recall seeing the copyright disclaimer last time. Is this okay with JOSS?
