### 2021-12-09 ORNL/tx2#1 @WildfireXIII
1. That's a good question, those don't show up when I locally compile it with JOSS's docker container, so I'm assuming that's just part of their draft mode?
2. Oh no haha, that might be a Fabian question, I don't manually insert reviewer names anywhere!
3. Definitely a Fabian question, I believe that section was there when we first submitted, it's in paper.md in the repository commit right before the submission date: https://github.com/ORNL/tx2/blob/60ff3367e3db5b3e0d5fba683d5b4fcb57b0ab33/paper/paper.md. I was told to add that bit I think, the software itself is definitely under a BSD license, but I'm definitely not a lawyer so I don't know what the paper license stuff implies! If this is a potential issue I can ask around about it. (I'll drop a note over in the main thread in the meantime)
