### 2021-12-09 #3652 @WildfireXIII
@fabian-s hey if you get a chance, could you drop by https://github.com/ORNL/tx2/issues/1? There's some questions out of my scope.
