### 2021-12-09 ORNL/tx2#1 @fabian-s
@1 I think so, too
@2 no clue. if this persists at the proofing stage i'll ask the veteran editors
@3 BSD-3 is fine, JOSS acceps all OSI-approved licenses
