### 2021-12-09 ORNL/tx2#1 @WildfireXIII
@fabian-s I think number 3 is referring to the legal jargon line in the acknowledgements section in the paper: https://github.com/ORNL/tx2/blob/master/paper/paper.md?plain=1#L59

@stewartsl do you know where we got that/is it required?
