### 2021-12-10 fbc3d35
Add cuda device override to wrapper and update docs
