+++ requirements.txt
@@ -1,4 +1,4 @@
-sklearn
+scikit-learn~=1.0
 nltk
 pandas
 numpy
