+++ README.md
@@ -26,14 +26,14 @@ pip install tx2
 NOTE: depending on the environment, it may be better to install some of the dependencies separately before
 pip installing tx2, e.g. in conda:
 ```bash
-conda install pytorch-gpu pandas scikit-learn matplotlib ipywidgets numpy -c
+conda install pytorch-gpu pandas scikit-learn~=1.0 matplotlib ipywidgets numpy -c
 conda-forge
 ```
 
 If you do not have access to a GPU on your machine, install the regular pytorch
 package:
 ```bash
-conda install pytorch pandas scikit-learn matplotlib ipywidgets numpy
+conda install pytorch pandas scikit-learn~=1.0 matplotlib ipywidgets numpy
 ```
 
 Note that `pytorch-gpu` can only be found in the `conda-forge` channel.
@@ -51,7 +51,7 @@ Packages you'll need to install for the notebooks to work (in addition to the
 conda installs above):
 
 ```bash
-pip install tqdm transformers>=4.1.1
+pip install tqdm transformers~=4.1.1
 ```
 
 Running through each full notebook will produce the ipywidget dashboard near the
