### 2021-12-15 d80c478
Revert get_feature_names to pre sklearn-1.0 due to versioning conflicts
