### 2021-12-15 3a803a8
Update setup and README for correct version numbers
