### 2021-12-20 #3652 @fabian-s
Wonderful, we might be able to get this done before the holidays. Thank you all for working on this submission.

@WildfireXIII  - final requested changes for style & spelling:

- [ ] please add missing DOIs to your references (see also bot above)
- [ ] please add the missing software references for `ipywidget` and `Sphinx`
- [ ] please correct formatting of references for "Black" (esp. in-line reference looks weird/nonstandard) and your own DOE code reference (DOIs appear twice)
- [ ]  l. 21: "have the same *challenge* of many deep network approaches in that it is *challenging*": repeated "challenge", alternative phrasing suggestion: "have the same *challenge* of many deep network approaches in that it is *difficult*"
- [ ] l. 90 "Sphnix" should be "Sphinx"
