### 2021-12-21 #3652 @fabian-s
Great, ty @WildfireXIII !

At this point could you:

I don't think your DOEcode entry counts as an "archival deposit" in our sense of the word since it just points to the GH repo in general, not a specific tagged commit in the repo.
