### 2021-12-21 a9dcc2a
Fix date in paper and re-run black formatter
