+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Oak Ridge National Laboratory
    index: 1
-date: 20 Decemeber 2021
+date: 21 December 2021
 bibliography: paper.bib
 ---
 
