+++ tx2/calc.py
@@ -140,7 +140,7 @@ def frequent_words_in_cluster(
     counter = CountVectorizer(stop_words=stopwords.words("english"))
     cv_fit = counter.fit_transform(texts)
     freq_words = sorted(
-        # NOTE: get_feature_names_out is scikit-learn 1.0, was having difficulty getting 
+        # NOTE: get_feature_names_out is scikit-learn 1.0, was having difficulty getting
         #   scikitlearn, numpy, and numba versions to play nice with that requirement however.
         #   for now leaving scikit-learn as pre-1.0
         zip(counter.get_feature_names(), cv_fit.toarray().sum(axis=0)),
@@ -189,7 +189,7 @@ def frequent_words_by_class_in_cluster(
             stop_words=stopwords.words("english"), vocabulary=vocab
         )
         cv_fit = counter.fit_transform(local_df.text.values)
-        # NOTE: get_feature_names_out is scikit-learn 1.0, was having difficulty getting 
+        # NOTE: get_feature_names_out is scikit-learn 1.0, was having difficulty getting
         #   scikitlearn, numpy, and numba versions to play nice with that requirement however.
         #   for now leaving scikit-learn as pre-1.0
         class_freq_words = list(
