### 2021-12-21 #3652 @WildfireXIII
Release tag: v1.0.0
Zenodo DOI: 10.5281/zenodo.5796089
